# docs/swagger.yaml
basePath: /v1
definitions:
...
  internal_app_auth.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  internal_app_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_app_auth.LoginResponse:
    properties:
      message:
        type: string
      refreshToken:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/internal_app_auth.UserResponse'
      valid:
        type: boolean
    type: object
  internal_app_auth.RefreshRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  internal_app_auth.ResetPasswordRequest:
    properties:
      email:
        type: string
      password:
        type: string
      token:
        type: string
    required:
    - email
    - password
    - token
    type: object
  internal_app_auth.UserResponse:
    properties:
      active:
        type: boolean
      avatar:
        type: string
      created_at:
        type: string
      document:
        type: string
      email:
        type: string
      id:
        type: string
      id_role:
        type: string
      name:
        type: string
      phone:
        type: string
      role: {}
      updated_at:
        type: string
    type: object
  internal_app_auth.ValidateTokenRequest:
    properties:
      email:
        type: string
      token:
        type: string
    required:
    - email
    - token
    type: object
...
host: localhost:8888
info:
  contact:
    email: suporte@swagger.io
    name: Suporte API
    url: http://www.swagger.io/support
  description: API modular em Go com Gin e Swagger.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend Go API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentica o usuário com email e senha, retornando tokens JWT e
        dados do usuário.
      parameters:
      - description: Credenciais de acesso
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_app_auth.LoginResponse'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Credenciais inválidas
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Realizar Login
      tags:
      - Auth
  /auth/me:
    get:
      description: Retorna informações do usuário logado baseado no token JWT.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_app_auth.LoginResponse'
        "401":
          description: Não autorizado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Obter dados do usuário atual
      tags:
      - Auth
  /auth/password/change:
    post:
      consumes:
      - application/json
      description: Altera a senha do usuário utilizando o código de validação.
      parameters:
      - description: Novos dados de senha
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_auth.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: senha alterada com sucesso'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Token inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Redefinir senha
      tags:
      - Auth
  /auth/password/request:
    post:
      consumes:
      - application/json
      description: Envia um e-mail com um código de 6 dígitos para o usuário.
      parameters:
      - description: Email do usuário
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: e-mail de recuperação enviado se o usuário existir'
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Solicitar recuperação de senha
      tags:
      - Auth
  /auth/password/validate:
    post:
      consumes:
      - application/json
      description: Verifica se o código de 6 dígitos é válido e não expirou.
      parameters:
      - description: Dados de validação
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_auth.ValidateTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'valid: true'
          schema:
            additionalProperties:
              type: boolean
            type: object
        "401":
          description: Token inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Validar código de recuperação
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Gera um novo access token a partir de um refresh token válido.
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_app_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_app_auth.LoginResponse'
        "400":
          description: Dados inválidos
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Refresh token inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Renovar token
      tags:
      - Auth
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"